add response carpools socket

+++ src/server.js
@@ -75,7 +75,7 @@ io.on('connect', client => {
     client.on("request", async req => {
         var status = 405
         const carpool_data = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
-            attributes: ["seats_total"] })
+            attributes: ["seats_total", "owner"] })
         const counts = await Request.findAll({ where: {
             carpool_id: req["carpool_id"],
             approved: true
@@ -89,26 +89,42 @@ io.on('connect', client => {
             passengers_count = 1
         }
         if(len <= carpool_data["seats_total"] - passengers_count) {
-            await Request
-                .findOrCreate({ where: {
-                    user_id: req["user_id"],
-                    carpool_id: req["carpool_id"]
-                } })
-                .then(([request]) => {
-                    if(passengers_count > 1) {
-                        Request.update({ count: passengers_count },
-                            { where: { id: request.id } })
+
+            var access = true
+
+            if(req["role"] == 2) {
+                for (const token in conn) {
+                    if(conn[token] == client.id) {
+                        if(token != carpool_data["owner"]) {
+                            access = false
+                        }
                     }
-                    if(request.author_role == null) {
-                        Request.update({ author_role: req["role"] },
-                            { where: { id: request.id } })
-                    }
-                })
-            status = 200
+                }
+            }
+
+            if(access) {
+                await Request
+                    .findOrCreate({ where: {
+                        user_id: req["user_id"],
+                        carpool_id: req["carpool_id"]
+                    } })
+                    .then(([request]) => {
+                        if(passengers_count > 1) {
+                            Request.update({ count: passengers_count },
+                                { where: { id: request.id } })
+                        }
+                        if(request.author_role == null) {
+                            Request.update({ author_role: req["role"] },
+                                { where: { id: request.id } })
+                        }
+                    })
+                status = 200
+            }
+
         }
 
         if(req["role"] == 1) {
-            const user = await User.findOne({ where: { token: req["user_id"] }, attributes: ["name"] })
+            const user = await User.findOne({ where: { token: req["user_id"] }, attributes: ["name", "token"] })
             const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
                 attributes: ["owner"] })
             if(conn[carpool_owner_token["owner"]] != null) {
@@ -123,33 +139,50 @@ io.on('connect', client => {
             const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
                 attributes: ["owner"] })
             const user = await User.findOne({ where: { token: carpool_owner_token["owner"] },
-                attributes: ["name"] })
+                attributes: ["name", "token"] })
             if(conn[req["user_id"]] != null) {
                 io.sockets.in(conn[req["user_id"]]).emit("request", {
                     peoples: passengers_count,
                     carpool_id: req["carpool_id"],
                     user: user,
-                    role: req["role"]
+                    role: req["role"],
+                    status: status
                 })
             }
         }
     })
-/*
+
     client.on("response", async req => {
         var status = req["status"]
-        if(status == 1) {
-            const request_data = await Request.findOne({ where: {
-                carpool_id: req["carpool_id"],
-                user_id: req["user_id"]
-            }, attributes: ["count", "author_role"] })
+        const request_data = await Request.findOne({ where: {
+            carpool_id: req["carpool_id"],
+            user_id: req["user_id"]
+        }, attributes: ["count", "author_role"] })
+        var role = request_data["author_role"]
+        const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
+            attributes: ["owner"] })
 
+        if(status == 1) {
             var access = true
-            if(request_data["author_role"] == 1) {
-                const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
-                attributes: ["owner"] })
-                if(carpool_owner_token["owner"] != req["token"]) {
-                    access = false
-                    status = 0
+            if(role == 1) {
+                for (const token in conn) {
+                    if(conn[token] == client.id) {
+                        if(token != carpool_owner_token["owner"]) {
+                            access = false
+                            status = 400
+                        }
+                    }
+                }
+            }
+
+            if(role == 2) {
+                for (const token in conn) {
+                    if(conn[token] == client.id) {
+                        if(token != req["user_id"]) {
+                            access = false
+                            status = 400
+                        }
+                    }
                 }
             }
 
@@ -170,42 +203,60 @@ io.on('connect', client => {
             }
 
         } else if(status == 2) {
-            await Request.destroy({ where: {
-                    user_id: req["user_id"],
-                    carpool_id: req["carpool_id"]
-                } 
-            })
+
+            var access = false
+            for (const token in conn) {
+                    if(conn[token] == client.id) {
+                        if(token == req["user_id"]) {
+                            access = true
+                            role = 2
+                        } else {
+                            if(token == carpool_owner_token["owner"]) {
+                                access = true
+                                role = 1
+                            } else {
+                                status = 400
+                            }
+                        }
+                    }
+                }
+
+            if(access) {
+                await Request.destroy({ where: {
+                        user_id: req["user_id"],
+                        carpool_id: req["carpool_id"]
+                    } 
+                })
+            }
         }
 
-        const role = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
-            attributes: ["author_role"] })
-        if(role["author_role"] == 1) {
-            const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
-                attributes: ["owner"] })
+        if(status == 400) {
+            client.emit("response", {
+                status: status
+            })
+        } else if(role == 1) {
             const user = await User.findOne({ where: { token: carpool_owner_token["owner"] },
                 attributes: ["name"] })
             if(conn[req["user_id"]] != null) {
-                io.sockets.in(conn[req["user_id"]]).emit("request", {
+                io.sockets.in(conn[req["user_id"]]).emit("response", {
                     status: status,
                     carpool_id: req["carpool_id"],
                     user: user,
-                    role: role["author_role"]
+                    role: role
                 })
             }
-        } else if(req["role"] == 2) {
+        } else if(role == 2) {
             const user = await User.findOne({ where: { token: req["user_id"] }, attributes: ["name"] })
-            const carpool_owner_token = await Carpool.findOne({ where: { carpool_id: req["carpool_id"] },
-                attributes: ["owner"] })
             if(conn[carpool_owner_token["owner"]] != null) {
-                io.sockets.in(conn[carpool_owner_token["owner"]]).emit("request", {
+                io.sockets.in(conn[carpool_owner_token["owner"]]).emit("response", {
                     status: status,
                     carpool_id: req["carpool_id"],
                     user: user,
-                    role: req["role"]
+                    role: role
                 })
             }
         }
-    })*/
+    })
 
     client.on("test", req => {
         if(conn[req["user_id"]] != null) {
